feat: オーバーフローチェックを追加 (#23)
* Heroku起動用のProcfileを追加

* fix: ProcfileでHeroku用に$PORTを使用するよう修正

* feat: amountの型をIntegerからLongに変更

* feat: 合計金額のオーバーフロー対策を追加

- 合計金額のオーバーフローを防ぐため、サービスクラスにチェック機能を追加。
- insert時の金額加算チェックを追加
- update時の差分に基づくオーバーフローチェックを追加

- ExpenseMapperに合計金額を取得するメソッドを追加。
- ExpenseMapper.xmlに対応するSQLを追加。

* feat: オーバーフロー時にエラーメッセージをフォーム画面に表示する処理を追加

- insert/update 時に発生する IllegalArgumentException をコントローラーで捕捉
- エラーメッセージをフォーム画面に渡す処理を追加
- form.html にエラーメッセージ表示領域を追加

+++ src/main/java/com/ozeken/expensecalendar/dto/DailyTotal.java
@@ -9,6 +9,6 @@ public class DailyTotal {
 	private int day;
 	
 	// 合計金額
-	private Integer total;
+	private Long total;
 	
 }
