+++ src/main/resources/schema.sql
@@ -18,7 +18,7 @@ CREATE TABLE IF NOT EXISTS expenses (
     user_id     INTEGER NOT NULL,                -- ユーザーID（外部キー）
     date        DATE NOT NULL,                       -- 日付
     genre_id    INTEGER NOT NULL,                -- ジャンルID（外部キー）
-    amount      INTEGER NOT NULL,               -- 金額
+    amount      BIGINT NOT NULL,               -- 金額
     description TEXT,                                       -- 説明
     CONSTRAINT fk_user FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE,
     CONSTRAINT fk_genre FOREIGN KEY (genre_id) REFERENCES genres(id)
